slack_bot/requirements.txt
# Slack SDK
slack-bolt>=1.18.0
slack-sdk>=3.23.0

# Web framework
flask>=3.0.0
flask-restx>=1.3.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
Werkzeug>=3.0.0

# Database
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0

# HTTP client
httpx>=0.25.0

# JWT authentication
PyJWT>=2.8.0

# Environment variables
python-dotenv>=1.0.0
